Gunnar: minor edit
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1698 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Hamiltonian/ParticleOnTorusCoulombWithMagneticTranslationsHamiltonian.h
@@ -72,7 +72,7 @@ class ParticleOnTorusCoulombWithMagneticTranslationsHamiltonian : public Abstrac
   // memory = maximum amount of memory that can be allocated for fast multiplication (negative if there is no limit)
   // precalculationFileName = option file name where precalculation can be read instead of reevaluting them
   ParticleOnTorusCoulombWithMagneticTranslationsHamiltonian(ParticleOnTorusWithMagneticTranslations* particles, int nbrParticles, int maxMomentum, int xMomentum,
-							    double ratio, bool haveCoulomb, int landauLevel, int nbrPseudopotentials, double* pseudopotentials, AbstractArchitecture* architecture, int memory = -1, char* precalculationFileName = 0);
+							    double ratio, bool haveCoulomb, int landauLevel, int nbrPseudopotentials, double* pseudopotentials, AbstractArchitecture* architecture, long memory = -1, char* precalculationFileName = 0);
 
   // destructor
   //
